% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_POVariable.R
\name{build_POVariable}
\alias{build_POVariable}
\title{build_POVariable}
\usage{
build_POVariable(.pathVar, .pathPO)
}
\arguments{
\item{.pathVar}{path Memory path where Variable Import month report is saved}

\item{.pathPO}{memory path where Standarized KOF Head Count Data base is located}
}
\value{
Standarized KOF Head Count Data base with added payer variable
}
\description{
build_POVariable
}
