% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_NominaVariable.R
\name{munge_NominaVariable}
\alias{munge_NominaVariable}
\title{munge_NominaVariable}
\usage{
munge_NominaVariable(.path)
}
\arguments{
\item{.path}{Memory path where Variable Import month report is saved}
}
\value{
.data dataframe with Employe ID and Total payed variable in MXN
}
\description{
munge_NominaVariable
}
